# ClientApp/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Proiect_disertatie",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:5001"
    }
  ],
  "paths": {
...
    "/api/Anime/GetAnimesFor/{username}": {
      "get": {
        "tags": [
          "Anime"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Anime/GetAllAnimes": {
      "get": {
        "tags": [
          "Anime"
        ],
        "parameters": [
          {
            "name": "SearchedAnime",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Anime/{title}": {
      "get": {
        "tags": [
          "Anime"
        ],
        "operationId": "GetAnime",
        "parameters": [
          {
            "name": "title",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Datum"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Datum"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Datum"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Anime/GetReviewsFor/{animeId}": {
      "get": {
        "tags": [
          "Anime"
        ],
        "parameters": [
          {
            "name": "animeId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Anime/AddAnimeToUser/{animeId}": {
      "post": {
        "tags": [
          "Anime"
        ],
        "parameters": [
          {
            "name": "animeId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Anime/AddReviewFor/{animeId}": {
      "post": {
        "tags": [
          "Anime"
        ],
        "parameters": [
          {
            "name": "animeId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Anime/DeleteAnimeFromUser/{animeId}": {
      "delete": {
        "tags": [
          "Anime"
        ],
        "parameters": [
          {
            "name": "animeId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Anime/DeleteReviewFor/{animeId}": {
      "delete": {
        "tags": [
          "Anime"
        ],
        "parameters": [
          {
            "name": "animeId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          {
            "name": "reviewId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Friends/CheckFriendship/{username}": {
      "get": {
        "tags": [
          "Friends"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/Friends/AcceptFriendRequest/{username}": {
      "post": {
        "tags": [
          "Friends"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Friends/SendFriendRequest/{username}": {
      "post": {
        "tags": [
          "Friends"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Friends/DeleteFriend/{username}": {
      "delete": {
        "tags": [
          "Friends"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Friends/CancelFriendRequest/{username}": {
      "delete": {
        "tags": [
          "Friends"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Friends/RefuseFriendRequest/{username}": {
      "delete": {
        "tags": [
          "Friends"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Aired": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "to": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "prop": {
            "$ref": "#/components/schemas/Prop"
          },
          "string": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Broadcast": {
        "type": "object",
        "properties": {
          "day": {
            "type": "string",
            "nullable": true
          },
          "time": {
            "type": "string",
            "nullable": true
          },
          "timezone": {
            "type": "string",
            "nullable": true
          },
          "string": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Datum": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "mal_id": {
            "type": "integer",
            "format": "int32"
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "images": {
            "$ref": "#/components/schemas/Images"
          },
          "trailer": {
            "$ref": "#/components/schemas/Trailer"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "title_english": {
            "type": "string",
            "nullable": true
          },
          "title_japanese": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "episodes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "airing": {
            "type": "boolean",
            "nullable": true
          },
          "aired": {
            "$ref": "#/components/schemas/Aired"
          },
          "duration": {
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "string",
            "nullable": true
          },
          "score": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "scored_by": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "rank": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "popularity": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "members": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "favorites": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "synopsis": {
            "type": "string",
            "nullable": true
          },
          "background": {
            "type": "string",
            "nullable": true
          },
          "season": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "broadcast": {
            "$ref": "#/components/schemas/Broadcast"
          },
          "producers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Producer"
            },
            "nullable": true
          },
          "licensors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Licensor"
            },
            "nullable": true
          },
          "studios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Studio"
            },
            "nullable": true
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Genre"
            },
            "nullable": true
          },
          "themes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Theme"
            },
            "nullable": true
          },
          "demographics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Demographic"
            },
            "nullable": true
          },
          "reviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Demographic": {
        "type": "object",
        "properties": {
          "mal_id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "From": {
        "type": "object",
        "properties": {
          "day": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "month": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "year": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Genre": {
        "type": "object",
        "properties": {
          "mal_id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Images": {
        "type": "object",
        "properties": {
          "jpg": {
            "$ref": "#/components/schemas/Jpg"
          },
          "webp": {
            "$ref": "#/components/schemas/Webp"
          },
          "image_url": {
            "type": "string",
            "nullable": true
          },
          "small_image_url": {
            "type": "string",
            "nullable": true
          },
          "medium_image_url": {
            "type": "string",
            "nullable": true
          },
          "large_image_url": {
            "type": "string",
            "nullable": true
          },
          "maximum_image_url": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Jpg": {
        "type": "object",
        "properties": {
          "image_url": {
            "type": "string",
            "nullable": true
          },
          "small_image_url": {
            "type": "string",
            "nullable": true
          },
          "large_image_url": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Licensor": {
        "type": "object",
        "properties": {
          "mal_id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ObjectId": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "machine": {
            "type": "integer",
            "format": "int32",
            "readOnly": true,
            "deprecated": true
          },
          "pid": {
            "type": "integer",
            "format": "int32",
            "readOnly": true,
            "deprecated": true
          },
          "increment": {
            "type": "integer",
            "format": "int32",
            "readOnly": true,
            "deprecated": true
          },
          "creationTime": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "Producer": {
        "type": "object",
        "properties": {
          "mal_id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Prop": {
        "type": "object",
        "properties": {
          "from": {
            "$ref": "#/components/schemas/From"
          },
          "to": {
            "$ref": "#/components/schemas/To"
          }
        },
        "additionalProperties": false
      },
...
      "ReviewDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true
          },
          "short_description": {
            "type": "string",
            "nullable": true
          },
          "main_description": {
            "type": "string",
            "nullable": true
          },
          "stars": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "Studio": {
        "type": "object",
        "properties": {
          "mal_id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Theme": {
        "type": "object",
        "properties": {
          "mal_id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "To": {
        "type": "object",
        "properties": {
          "day": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "month": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "year": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Trailer": {
        "type": "object",
        "properties": {
          "youtube_id": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "embed_url": {
            "type": "string",
            "nullable": true
          },
          "images": {
            "$ref": "#/components/schemas/Images"
          }
        },
        "additionalProperties": false
      },
...
      "Webp": {
        "type": "object",
        "properties": {
          "image_url": {
            "type": "string",
            "nullable": true
          },
          "small_image_url": {
            "type": "string",
            "nullable": true
          },
          "large_image_url": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}